git-svn-id: https://pcsxr.svn.codeplex.com/svn/pcsxr@67247 e17a0e51-4ae3-4d35-97c3-1a29b211df97

+++ libpcsxcore/cdrom.c
@@ -2451,7 +2451,7 @@ void cdrReset() {
 }
 
 int cdrFreeze(gzFile f, int Mode) {
-	uintptr_t tmp;
+	unsigned int tmp;
 
 
 	if( Mode == 0 ) {
